store-server/store/requirements.txt
Django==3.2.17
Pillow==9.4.0

django-allauth==0.52.0
django-environ==0.10.0
django-extensions==3.2.1
django_debug_toolbar==3.8.1

flake8==6.0.0
isort==5.12.0

psycopg2-binary==2.9.5

celery==5.2.7
django-redis==5.2.0
redis==4.5.1

yookassa==2.3.5

djangorestframework~=3.14.0